--- public/models/Detectors/components/DetectorBasicDetailsView/DetectorBasicDetailsView.mock.ts
@@ -1,9 +1,0 @@
-import { mockDetector } from '../../../Interfaces.mock';
-
-export default {
-  detector: mockDetector,
-  rulesCanFold: true,
-  enabled_time: 1,
-  last_update_time: 1,
-  onEditClicked: () => jest.fn(),
-};
