--- public/models/Detectors/components/DetectorRulesView/DetectorRulesView.mock.ts
@@ -1,8 +1,0 @@
-import { mockDetector, mockNotificationsStart } from '../../../Interfaces.mock';
-
-export default {
-  detector: mockDetector,
-  rulesCanFold: false,
-  onEditClicked: jest.fn(),
-  notifications: mockNotificationsStart,
-};
