--- public/models/Detectors/components/FieldMappingsView/FieldMappingsView.mock.ts
@@ -1,13 +1,0 @@
-import { FieldMapping } from '../../../../../models/interfaces';
-import { mockDetector, mockNotificationsStart } from '../../../Interfaces.mock';
-
-const mockFieldMapping: FieldMapping = {
-  indexFieldName: 'indexFieldName',
-  ruleFieldName: 'ruleFieldName',
-};
-export default {
-  detector: mockDetector,
-  existingMappings: [mockFieldMapping, mockFieldMapping],
-  editFieldMappings: jest.fn(),
-  notifications: mockNotificationsStart,
-};
