--- public/models/Detectors/components/UpdateDetectorBasicDetails/UpdateDetectorBasicDetails.mock.ts
@@ -1,10 +1,0 @@
-import { mockDetectorHit, mockHistory, mockNotificationsStart } from '../../../Interfaces.mock';
-
-export default {
-  notifications: mockNotificationsStart,
-  detectorHit: mockDetectorHit,
-  location: {
-    pathname: '/edit-detector-details/detectorHitId',
-  },
-  history: mockHistory,
-};
