--- public/models/Detectors/components/UpdateDetectorRules/UpdateDetectorRules.mock.ts
@@ -1,14 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { mockDetectorHit, mockNotificationsStart } from '../../../Interfaces.mock';
-
-export default {
-  notifications: mockNotificationsStart,
-  detectorHit: mockDetectorHit,
-  location: {
-    pathname: '',
-  },
-};
