--- public/models/Detectors/containers/AlertTriggersView/AlertTriggersView.mock.ts
@@ -1,7 +1,0 @@
-import { mockDetector, mockNotificationsStart } from '../../../Interfaces.mock';
-
-export default {
-  detector: mockDetector,
-  editAlertTriggers: jest.fn(),
-  notifications: mockNotificationsStart,
-};
