--- public/models/Detectors/containers/Detectors/Detectors.mock.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { mockDetectorService, mockNotificationsStart } from '../../../Interfaces.mock';
-
-export default {
-  detectorService: mockDetectorService,
-  notifications: mockNotificationsStart,
-};
