+++ public/pages/Detectors/components/FieldMappingsView/FieldMappingsView.test.tsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import props from '../../../../models/Detectors/components/FieldMappingsView/FieldMappingsView.mock';
+import props from '../../../../../test/mocks/Detectors/components/FieldMappingsView/FieldMappingsView.mock';
 import { expect } from '@jest/globals';
 import { FieldMappingsView } from './FieldMappingsView';
