+++ public/pages/Detectors/containers/DetectorDetailsView/DetectorDetailsView.test.tsx
@@ -4,7 +4,7 @@
  */
 
 import React from 'react';
-import props from '../../../../models/Detectors/containers/DetectorDetailsView/DetectorDetailsView.mock';
+import props from '../../../../../test/mocks/Detectors/containers/DetectorDetailsView/DetectorDetailsView.mock';
 import { expect } from '@jest/globals';
 import { act } from 'react-dom/test-utils';
 import { mount } from 'enzyme';
