+++ test/mocks/Detectors/components/FieldMappingsView/FieldMappingsView.mock.ts
@@ -1,0 +1,15 @@
+import { FieldMapping } from '../../../../../models/interfaces';
+import { mockDetector } from '../../containers/Detectors/Detectors.mock';
+import { mockNotificationsStart } from '../../../browserServicesMock';
+
+const mockFieldMapping: FieldMapping = {
+  indexFieldName: 'indexFieldName',
+  ruleFieldName: 'ruleFieldName',
+};
+
+export default {
+  detector: mockDetector,
+  existingMappings: [mockFieldMapping, mockFieldMapping],
+  editFieldMappings: jest.fn(),
+  notifications: mockNotificationsStart,
+};
