+++ test/mocks/Detectors/components/UpdateDetectorRules/UpdateDetectorRules.mock.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { mockDetectorHit } from '../../containers/Detectors/Detectors.mock';
+import { mockNotificationsStart } from '../../../browserServicesMock';
+
+export default {
+  notifications: mockNotificationsStart,
+  detectorHit: mockDetectorHit,
+  location: {
+    pathname: '',
+  },
+};
