+++ test/mocks/index.ts
@@ -6,4 +6,8 @@
 import browserServicesMock from './browserServicesMock';
 import httpClientMock from './httpClientMock';
 
+export const mockHistory = {
+  replace: jest.fn(),
+};
+
 export { browserServicesMock, httpClientMock };
